TWITTER SENTIMENTAL ANALYSIS USING HYBRID APPROACH
Sentimental Analysis is the process of determining the opinion of the writer based on their messages on various social networking sites. Twitter is one of the famous social networking sites where user can read and post messages about a person, an event, a product and the current happenings all over the world. These are normally 140-280 characters in length. In this system, the tweets are used as the raw data. The tweets are collected through Twitter API using the secret tokens. Then they are preprocessed using text mining package to reduce the noise in words. The score is computed for each pre-processed tweet using Dictionary-Based Approach. For positive tweet the score is 1, for negative tweet the score is -1 and 0 for neutral tweet. The pre-processed tweets along with the scores are stored in CSV format for further process. The train data and test data is provided in the ratio 60:40 to construct the classification model. After classification, Convolutional Neural Network is implemented to compute the probability of the tweets. The system uses K-fold cross validation method to improve over the holdout method. Finally, as the result the opinion of the sentiment related to the given tweets is predicted using probability of the positive tweets by hybrid approach. This system produces a better performance measure when compared to other method.